UPBGE: Fix MSVC 17.7 prev 5 build error
a build error with Visual studio 17.7 preview 5

+++ intern/ghost/intern/GHOST_Wintab.cc
@@ -188,7 +188,7 @@ GHOST_Wintab::GHOST_Wintab(unique_hmodule handle,
       m_context{std::move(hctx)},
       m_tabletCoord{tablet},
       m_systemCoord{system},
-      m_pkts{queueSize}
+      m_pkts{size_t(queueSize)}
 {
   m_fpInfo(WTI_INTERFACE, IFC_NDEVICES, &m_numDevices);
   WINTAB_PRINTF("Wintab Devices: %d\n", m_numDevices);
